Use long max_attempts to avoid casting, feed directly to AWS API

+++ cpp/src/arrow/filesystem/s3fs.cc
@@ -243,17 +243,15 @@ class AwsRetryStrategy : public S3RetryStrategy {
 };
 
 std::shared_ptr<S3RetryStrategy> S3RetryStrategy::GetAwsDefaultRetryStrategy(
-    int64_t max_attempts) {
+    long max_attempts) {
   return std::make_shared<AwsRetryStrategy>(
-      std::make_shared<Aws::Client::DefaultRetryStrategy>(
-          static_cast<long>(max_attempts)));
+      std::make_shared<Aws::Client::DefaultRetryStrategy>(max_attempts));
 }
 
 std::shared_ptr<S3RetryStrategy> S3RetryStrategy::GetAwsStandardRetryStrategy(
-    int64_t max_attempts) {
+    long max_attempts) {
   return std::make_shared<AwsRetryStrategy>(
-      std::make_shared<Aws::Client::StandardRetryStrategy>(
-          static_cast<long>(max_attempts)));
+      std::make_shared<Aws::Client::StandardRetryStrategy>(max_attempts));
 }
 
 // -----------------------------------------------------------------------
